app/requirements.txt
celery>=5.2.0
redis>=4.5.0
fastapi>=0.95.0
uvicorn[standard]>=0.23.0
numpy>=1.24.0
pillow>=9.5.0
torch>=2.0.0
transformers>=4.40.0
faiss-cpu>=1.7.4
tenseal>=0.6.1
cryptography>=41.0.0
streamlit>=1.25.0
pytest>=8.0.0
pytest-asyncio>=1.0.0
